#23 Move part of data into static data source. Add some of the missing localizations.
