+++ Oblasti UA/Controllers/SettingsViewController+NavigationController/en.lproj/SettingsViewController+NavigationController.strings
@@ -1,6 +1,6 @@
 
-/* Class = "UILabel"; text = "Restore default settings"; ObjectID = "Ob3-Oo-D7c"; */
-"Ob3-Oo-D7c.text" = "Restore default settings";
+/* Class = "UILabel"; text = "Restore default settings"; ObjectID = "NKk-yi-WN3"; */
+"NKk-yi-WN3.text" = "Restore default settings";
 
 /* Class = "UILabel"; text = "Classic"; ObjectID = "PgA-YM-8w5"; */
 "PgA-YM-8w5.text" = "Classic";
