+++ Oblasti UA/Controllers/SettingsViewController+NavigationController/ru.lproj/SettingsViewController+NavigationController.strings
@@ -1,6 +1,6 @@
 
-/* Class = "UILabel"; text = "Restore default settings"; ObjectID = "Ob3-Oo-D7c"; */
-"Ob3-Oo-D7c.text" = "Восстановить значения по умолчанию";
+/* Class = "UILabel"; text = "Restore default settings"; ObjectID = "NKk-yi-WN3"; */
+"NKk-yi-WN3.text" = "Восстановить значения по умолчанию";
 
 /* Class = "UILabel"; text = "Classic"; ObjectID = "PgA-YM-8w5"; */
 "PgA-YM-8w5.text" = "Классический";
